requirements.txt
fastapi==0.115.12
uvicorn==0.34.2
pydantic==2.11.4
python-dotenv==1.1.0
manim==0.19.0
boto3==1.26.0
google-genai==1.13.0
streamlit==1.32.0
gunicorn==21.2.0 
